+++ tests/functional/test_run.py
@@ -1,6 +1,7 @@
 import json
 import shutil
 import subprocess
+import bluepysnap
 import tempfile
 from pathlib import Path
 from subprocess import CalledProcessError
@@ -9,7 +10,7 @@ import h5py
 import pytest
 from click.testing import CliRunner
 from utils import SNAKEFILE, SNAKEMAKE_ARGS, TEST_PROJ_TINY, cwd, edit_yaml, load_yaml
-
+from assertions import assert_node_population_morphologies_accessible
 from circuit_build.cli import run
 
 
@@ -75,6 +76,12 @@ def test_functional_all(tmp_path):
                 config["networks"]["edges"][0]["edges_file"]
                 == f"$BASE_DIR/networks/edges/functional/{edge_population_name}/edges.h5"
             )
+
+        assert_node_population_morphologies_accessible(
+            circuit=bluepysnap.Circuit(tmp_path / "sonata/circuit_config.json"),
+            population_name=node_population_name,
+            extensions=["asc", "h5"],
+        )
 
 
 def test_no_emodel(tmp_path):